fix(sw.js): Fix suggestion #1. Caching optimising.
Category: Application Data and Offline Use

Reviewers Suggestion:
Suggestion
Caching resources dynamically or caching URLs during install event works but remember this is not a perfect solution. Think of the situation when the size grows up and/or has lot of URLs with dynamic query parameters !

Remember the role of service worker is to serve the resource.
Since we have all the resources required to load any detail page( restaurant.html & js files) there is no reason why we cannot load any details page offline , even if the user hasn't visited it earlier.

Hint
Please take a look at the URL in Restaurant Info screen, it's http://localhost:8000/restaurant.html?id=1 so even though you stored restaurant.html in the cache but it still hasn't found this in the cache, you should remove the query behind the /restaurant.html and it will work with every restaurant.

Check out the ignoreSearch in the external resource below to get it working

External Resources :books:
https://developer.mozilla.org/en-US/docs/Web/API/Cache/match