feat(*): Add UI and functionality for favorite selection and presentation
Add select dropdown fill functionality and filtering on map
Add and styled both index.html and restaurant.html with a star svg image
in two colors for marking a favorited restaurant and a not favorited
restaurant.

Pends the onclick event to change the star svg and also update the db.